+++ apps/backend/src/app/database/services/coding/index.ts
@@ -12,7 +12,8 @@ export { CodingReplayAnchorService } from './coding-replay-anchor.service';
 export { CodingReplayService } from './coding-replay.service';
 export { CodingResponseQueryService } from './coding-response-query.service';
 export { CodingResultsService } from './coding-results.service';
-export { CodingReviewService } from './coding-review.service';
+export { DoubleCodingReviewQueryService } from './double-coding-review-query.service';
+export { DoubleCodingReviewDecisionService } from './double-coding-review-decision.service';
 export { CodingStatisticsService } from './coding-statistics.service';
 export { CodingReadinessService } from './coding-readiness.service';
 export { CodingValidationService } from './coding-validation.service';
