+++ apps/backend/src/app/workspace/workspace.module.ts
@@ -74,7 +74,8 @@ import {
   CodingListExportService,
   CodingExportOrchestratorService,
   CodingValidationService,
-  CodingReviewService,
+  DoubleCodingReviewDecisionService,
+  DoubleCodingReviewQueryService,
   CodingAnalysisService,
   CodingProgressService,
   CodingReplayService,
@@ -174,7 +175,8 @@ import { CodingModule } from '../coding/coding.module';
     CodingListExportService,
     CodingExportOrchestratorService,
     CodingValidationService,
-    CodingReviewService,
+    DoubleCodingReviewDecisionService,
+    DoubleCodingReviewQueryService,
     CodingAnalysisService,
     CodingProgressService,
     CodingReplayService,
@@ -218,7 +220,8 @@ import { CodingModule } from '../coding/coding.module';
     CodingListExportService,
     CodingExportOrchestratorService,
     CodingValidationService,
-    CodingReviewService,
+    DoubleCodingReviewDecisionService,
+    DoubleCodingReviewQueryService,
     CodingAnalysisService,
     CodingProgressService,
     CodingReplayService,
